requirements.txt
# Portugal Data Intelligence - Python Dependencies
# ================================================
# Install with: pip install -r requirements.txt

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# API Data Fetching
requests>=2.31.0,<3.0.0

# Database
# sqlite3 is included in the Python standard library (no installation required)

# Visualisation
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.15.0,<6.0.0

# Statistical Analysis & Machine Learning
scipy>=1.10.0,<2.0.0
scikit-learn>=1.3.0,<2.0.0
statsmodels>=0.14.0,<1.0.0

# Report Generation
python-pptx>=0.6.21,<1.0.0

# Jupyter Notebooks
jupyter>=1.0.0
notebook>=7.0.0

# AI Insights
openai>=1.0.0,<2.0.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.65.0
